+++ src/components/global-types.ts
@@ -14,6 +14,12 @@ export enum InputSize {
     LARGE = "large",
 }
 
+export enum ButtonSize {
+    SMALL = "small",
+    MEDIUM = "medium",
+    LARGE = "large",
+}
+
 export type FloatingInputWidth = LabelPercentageWidth | 91 | 92 | 93 | 94 | 95 | 96 | 97 | 98 | 99 | 100;
 
 export type LabelPercentageWidth =